validations-epam-email/src/main/java/com/epam/rd/autotasks/validations/EpamEmailValidation.java
package com.epam.rd.autotasks.validations;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

public class EpamEmailValidation {

    public static boolean validateEpamEmail(String email) {
        // Put your code here
        if (email == null || email.isBlank()) {
            return false;
        }
        String regExp = "(\\w+)_(\\w+)(@epam.com)";
        Pattern pattern = Pattern.compile(regExp, Pattern.CASE_INSENSITIVE);
        Matcher matcher = pattern.matcher(email);
        return matcher.matches();

    }
}




